Move SimpleHDF5Mapping to separate file.
